pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tapegif-pro"
version = "1.0.0"
description = "tapegif Pro — TUI recorder, WebP/APNG output, watermarks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["gif", "webp", "apng", "terminal", "tui", "recorder", "animation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "tapegif>=0.1.0",
    "textual>=0.80.0",
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
    "click>=8.0.0",
]

[project.scripts]
tapegif-pro = "tapegif_pro.cli:main"

[project.urls]
Homepage = "https://github.com/Gnomecromancer/tapegif-pro"
Repository = "https://github.com/Gnomecromancer/tapegif-pro"

[tool.setuptools.packages.find]
where = ["."]
include = ["tapegif_pro*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
